requirements.txt
psycopg2==2.9.5
psycopg2_binary==2.9.5
python-decouple==3.6
